+++ NEWS.md
@@ -5,3 +5,4 @@
 * Added branch_name and git_pull arguments to the load_pkgs function. These arguments allow the user to specify a specific git branch to load packages from, and to pull the latest changes from the repository before loading the packages.
 * Enhance add_pkgs to check if the path is an R package directory before adding to multiloadr.
 * Enhance load_pkgs to check if the directory has a remote URL when git_pull = TRUE
+* Add from_commit argument to load package from specific commit